refactor(api): one class per middleware file
CLAUDE.md sacred rule #1 ("one class per file, always") was violated by
api/middleware.py holding both RequestIdMiddleware and AccessLogMiddleware.
Split each into its own file (request_id_middleware.py, access_log_middleware.py)
and reduce middleware.py to the configure_middleware() stack-assembly function.

Tests that imported the classes from app.api.middleware now import directly
from the split files — no backwards-compat shim.

No behavior change: configure_middleware still adds middleware in the same
outside-in order (CORS, RequestId, AccessLog).
